html/browsers/the-window-object/window-named-properties.html

<!doctype html>
<meta charset=utf-8>
<title>Changes to named properties of the window object</title>
<link rel="author" title="Ms2ger" href="ms2ger@gmail.com">
<link rel="author" title="Boris Zbarsky" href="bzbarsky@mit.edu">
<link rel="help" href="http://www.whatwg.org/html/#window">
<link rel="help" href="http://www.whatwg.org/html/#dom-window-nameditem">
<link rel="help" href="http://dev.w3.org/2006/webapi/WebIDL/#named-properties-object">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<iframe name="bar"></iframe>
<iframe name="constructor"></iframe>
<script>
function assert_data_propdesc(pd, Writable, Enumerable, Configurable) {
  assert_equals(typeof pd, "object");
  assert_equals(pd.writable, Writable, "Writable");
  assert_equals(pd.enumerable, Enumerable, "Enumerable");
  assert_equals(pd.configurable, Configurable, "Configurable");
}

var expected;
setup(function() {
  expected = [
    ["bar", document.getElementsByTagName("iframe")[0].contentWindow],
  ];
  ["a", "applet", "area", "embed", "form", "frameset", "img", "object"].forEach(function(localName) {
    var element = document.createElement(localName);
    var name = "name-" + localName;
    element.setAttribute("name", name);
    document.body.appendChild(element);
    expected.push([name, element]);
  });
});

expected.forEach(function(e) {
  var name = e[0], object = e[1];
  test(function() {
    assert_true(name in window, "name not in window");
    assert_equals(window[name], object);
    assert_false(window.hasOwnProperty(name), "window.hasOwnProperty(name)");
  }, "Static name " + format_value(name) + " on window");

  test(function() {
    assert_true(name in Window.prototype, "name not in Window.prototype");
    assert_equals(Window.prototype[name], object);
    assert_false(Window.prototype.hasOwnProperty(name), "Window.prototype.hasOwnProperty(name)");
  }, "Static name " + format_value(name) + " on Window.prototype");

  test(function() {
    var gsp = Object.getPrototypeOf(Object.getPrototypeOf(window));
    assert_true(name in gsp, "name not in gsp");
    assert_equals(gsp[name], object);
    assert_true(gsp.hasOwnProperty(name), "Window.prototype.hasOwnProperty(name)");
    assert_data_propdesc(Object.getOwnPropertyDescriptor(gsp, name),
                         false, true, true);
  }, "Static name " + format_value(name) + " on the named properties object");
});

test(function() {
  assert_equals(window.constructor, Window);
  assert_false(window.hasOwnProperty("constructor"), "window.constructor should not be an own property.");

  var proto = Object.getPrototypeOf(window);
  assert_equals(proto.constructor, Window);
  assert_true("constructor" in proto, "constructor in proto");
  assert_data_propdesc(Object.getOwnPropertyDescriptor(proto, "constructor"),
                       true, false, true);

  var gsp = Object.getPrototypeOf(proto);
  assert_true("constructor" in gsp, "constructor in gsp");
  assert_true(gsp.hasOwnProperty("constructor"), "gsp.hasOwnProperty(\"constructor\")");
  assert_data_propdesc(Object.getOwnPropertyDescriptor(gsp, "constructor"),
                       false, true, true);
}, "constructor");
var t = async_test("Dynamic name")
var t2 = async_test("Ghost name")
t.step(function() {
  var iframe = document.getElementsByTagName("iframe")[0];
  iframe.setAttribute("src", "data:text/html,<script>window.name='foo'<\/script>");
  iframe.onload = function() {
    t.step(function() {
      assert_true("foo" in window, "foo not in window");
      assert_equals(window["foo"], iframe.contentWindow);
    });
    t.done();
    t2.step(function() {
      assert_false("bar" in window, "bar still in window");
      assert_equals(window["bar"], undefined);
    });
    t2.done();
  };
});
</script>
