third_party/WebKit/LayoutTests/imported/web-platform-tests/html/browsers/the-window-object/window-named-properties-expected.txt

This is a testharness.js-based test.
PASS Static name 
FAIL Static name on the prototype assert_false: Window.prototype.hasOwnProperty("bar") expected false got true
FAIL constructor assert_equals: expected false but got true
PASS Dynamic name 
PASS Ghost name 
Harness: the test ran to completion.
